reload bug fixed
